Moving file delete controllers out of upload.py
Upload is getting long and delete doesn't depend on anything in upload.py
